+++ .gitignore
@@ -1,0 +1,2 @@
+.vscode
+.pynb_checkpoints
